refactor: clarify parser and renderer boundaries
